wsutil: More -Wsign-conversion issues
Some of things are casting inherently unsigned objects that are
signed in POSIX for historical reasons (e.g., blksize_t). The
number of arguments when spawning a command in ws_pipe also must
be unsigned. g_get_monotonic_time returns a signed 64-bit integer
for some reason.

+++ wsutil/file_compressed.c
@@ -187,7 +187,7 @@ writecap_file_open(ws_cwstream* pfile, const char *filename)
 
                 if (ws_stat64(filename, &statb) == 0) {
                     if (statb.st_blksize > IO_BUF_SIZE) {
-                        buffsize = statb.st_blksize;
+                        buffsize = (size_t)statb.st_blksize;
                     }
                 }
 #endif
@@ -224,7 +224,7 @@ writecap_file_fdopen(ws_cwstream* pfile, int fd)
 
                 if (ws_fstat64(fd, &statb) == 0) {
                     if (statb.st_blksize > IO_BUF_SIZE) {
-                        buffsize = statb.st_blksize;
+                        buffsize = (size_t)statb.st_blksize;
                     }
                 }
 #endif
